+++ e2e/jest/nestjs/e2e-assets.ts
@@ -1,7 +1,4 @@
-import 'reflect-metadata';
 import { forwardRef, Inject, Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
 
 @Injectable()
 export class TestClassOne {
@@ -35,29 +32,14 @@ export class TestClassFour {
   }
 }
 
+export interface Repository<T> {
+  value: T;
+}
+
 export class ClassThatIsNotInjected {}
 
 export interface Logger {
   log(): any;
-}
-
-@Injectable()
-export class MainTestClass {
-  constructor(
-    private readonly testClassOne: TestClassOne,
-    private readonly testClassTwo: TestClassTwo,
-    private readonly testClassTwoSecond: TestClassTwo,
-    @Inject('LOGGER') private readonly logger: Logger,
-    @Inject('PRIMITIVE_VALUE') private readonly primitiveValue: string
-  ) {}
-
-  async test(): Promise<string> {
-    const value = await this.testClassOne.foo(true);
-    const value2 = await this.testClassTwo.bar();
-    const value3 = this.logger.log();
-
-    return `${value}-${value2}-${value3}-${this.primitiveValue}`;
-  }
 }
 
 export class Foo {}
@@ -67,12 +49,6 @@ type Relation<T> = T;
 
 @Injectable()
 export class NestJSTestClass {
-  @Inject(TestClassOne)
-  private readonly testClassOne: TestClassOne;
-
-  @InjectRepository(Bar)
-  private readonly barRepository: Repository<Bar>;
-
   public constructor(
     @Inject('LOGGER') private readonly logger: Logger,
     @Inject('UNDEFINED') private readonly undefinedParam: undefined,
@@ -80,16 +56,53 @@ export class NestJSTestClass {
     @Inject(TestClassFour) private readonly testClassFour: undefined,
     @Inject(forwardRef(() => TestClassThree))
     private readonly testClassThree: Relation<TestClassThree>,
-    @InjectRepository(Foo) private readonly fooRepository: Repository<Foo>,
+    @Inject(Foo) private readonly fooRepository: Repository<Foo>,
     private readonly testClassTwo: TestClassTwo,
-    @Inject('PRIMITIVE_VALUE') private readonly primitiveValue: string
+    @Inject('PRIMITIVE_VALUE') private readonly primitiveValue: string,
+    private readonly testClassOne: TestClassOne
   ) {}
 
   async test(): Promise<string> {
-    const value = await this.testClassOne.foo(true);
     const value2 = await this.testClassTwo.bar();
     const value3 = this.logger.log();
 
-    return `${value}-${value2}-${value3}`;
+    return `${value2}-${value3}`;
+  }
+}
+
+@Injectable()
+export class NestJSTestClassProp {
+  @Inject('LOGGER')
+  private readonly logger: Logger;
+
+  @Inject('UNDEFINED')
+  private readonly undefinedParam: undefined;
+
+  @Inject('UNDEFINED_SECOND')
+  private readonly undefinedParamSecond: undefined;
+
+  @Inject(TestClassFour)
+  private readonly testClassFour: undefined;
+
+  @Inject(forwardRef(() => TestClassThree))
+  private readonly testClassThree: Relation<TestClassThree>;
+
+  @Inject(Foo)
+  private readonly fooRepository: Repository<Foo>;
+
+  @Inject(TestClassTwo)
+  private readonly testClassTwo: TestClassTwo;
+
+  @Inject('PRIMITIVE_VALUE')
+  private readonly primitiveValue: string;
+
+  @Inject(TestClassOne)
+  private readonly testClassOne: TestClassOne;
+
+  async test(): Promise<string> {
+    const value2 = await this.testClassTwo.bar();
+    const value3 = this.logger.log();
+
+    return `${value2}-${value3}`;
   }
 }